backend/requirements.txt
flask==2.0.1
werkzeug==2.0.3
sqlalchemy==1.4.23
flask-sqlalchemy==2.5.1
flask-jwt-extended==4.3.1
flask-cors==3.0.10
python-dotenv==0.19.0
requests==2.26.0
gunicorn==20.1.0 